LayoutTests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: line 10: TypeError: Type error
